Generate Algolia search index from nuxt/dist after full build
