+++ lib/search-index.js
@@ -271,8 +271,97 @@ function extractHeadingRecords({
     return records;
 }
 
+function extractMetaTag(html, selector) {
+    const regex = new RegExp(`<meta[^>]*${selector}[^>]*content=(["'])(.*?)\\1[^>]*>`, "i");
+    const match = html.match(regex);
+    return match?.[2] || "";
+}
+
+function extractHtmlTitle(html) {
+    const match = html.match(/<title[^>]*>([\s\S]*?)<\/title>/i);
+    return match ? match[1].replace(/\s+/g, " ").trim() : "";
+}
+
+function normalizeImage(value, fallback) {
+    if (!value) return fallback;
+    if (typeof value === "string") return value;
+    if (typeof value === "object" && typeof value.src === "string") return value.src;
+    return fallback;
+}
+
+function extractSearchHtml(url, html = "") {
+    const mainMatch = html.match(/<main[^>]*>([\s\S]*?)<\/main>/i);
+    let searchable = mainMatch ? mainMatch[1] : html;
+    if (url.startsWith("/blog/")) {
+        searchable = searchable.split("<!-- Author Bio Section -->")[0];
+    }
+    return searchable;
+}
+
+function toUnixTimestampSeconds(value) {
+    if (!value) return null;
+    const date = value instanceof Date ? value : new Date(value);
+    if (Number.isNaN(date.getTime())) return null;
+    return Math.floor(date.getTime() / 1000);
+}
+
+async function listHtmlFiles(rootDir) {
+    const fs = require("fs");
+    const path = require("path");
+    const files = [];
+    async function walk(currentDir) {
+        const entries = await fs.promises.readdir(currentDir, { withFileTypes: true });
+        for (const entry of entries) {
+            const fullPath = path.join(currentDir, entry.name);
+            if (entry.isDirectory()) {
+                await walk(fullPath);
+            } else if (entry.isFile() && entry.name.endsWith(".html")) {
+                files.push(fullPath);
+            }
+        }
+    }
+    await walk(rootDir);
+    return files;
+}
+
+function outputPathToUrl(outputRoot, outputPath) {
+    const path = require("path");
+    const rel = path.relative(outputRoot, outputPath).replace(/\\/g, "/");
+    if (!rel.endsWith(".html")) return "";
+    if (rel === "index.html") return "/";
+    if (rel.endsWith("/index.html")) return `/${rel.slice(0, -"index.html".length)}`;
+    return `/${rel}`;
+}
+
+function extractMetaKeywords(html) {
+    const raw = extractMetaTag(html, 'name=["\\\']keywords["\\\']');
+    if (!raw) return [];
+    return raw.split(",").map((keyword) => decodeEntities(keyword.trim()));
+}
+
+function extractDateFromJsonLd(html, fieldName) {
+    const regex = new RegExp(`"${fieldName}"\\s*:\\s*"([^"]+)"`, "i");
+    const match = html.match(regex);
+    return match?.[1] || "";
+}
+
+function extractArticleSection(html) {
+    return extractMetaTag(html, 'property=["\\\']article:section["\\\']').trim();
+}
+
 module.exports = {
     isSearchPage,
     isSearchUrl,
     extractHeadingRecords,
+    decodeEntities,
+    extractMetaTag,
+    extractHtmlTitle,
+    normalizeImage,
+    extractSearchHtml,
+    toUnixTimestampSeconds,
+    listHtmlFiles,
+    outputPathToUrl,
+    extractMetaKeywords,
+    extractDateFromJsonLd,
+    extractArticleSection,
 };
